requirements.txt
PyYAML>=5.3.1,<6
rfeed>=1.1.1,<2
